// generated-types/models/UserNotificationDto.ts
export interface UserNotificationDto {
    /**
     * Notification ID
     * @type {number}
     * @memberof UserNotificationDto
     */
    id?: number;
    /**
     * ID of the user this notification belongs to
     * @type {number}
     * @memberof UserNotificationDto
     */
    userId?: number;
    /**
     * Notification message content
     * @type {string}
     * @memberof UserNotificationDto
     */
    message?: string;
    /**
     * Type of notification
     * @type {string}
     * @memberof UserNotificationDto
     */
    type?: UserNotificationDtoTypeEnum;
    /**
     * Read/unread status of the notification
     * @type {string}
     * @memberof UserNotificationDto
     */
    status?: UserNotificationDtoStatusEnum;
    /**
     * Result of the notification action
     * @type {string}
     * @memberof UserNotificationDto
     */
    result?: UserNotificationDtoResultEnum;
    /**
     * Timestamp when the notification was created
     * @type {Date}
     * @memberof UserNotificationDto
     */
    createdAt?: Date;
}


/**
 * @export
 */
export const UserNotificationDtoTypeEnum = {
    ProgressUploadSuccess: 'PROGRESS_UPLOAD_SUCCESS',
    ProgressUploadFailure: 'PROGRESS_UPLOAD_FAILURE'
} as const;
export type UserNotificationDtoTypeEnum = typeof UserNotificationDtoTypeEnum[keyof typeof UserNotificationDtoTypeEnum];

/**
 * @export
 */
export const UserNotificationDtoStatusEnum = {
    Unread: 'UNREAD',
    Read: 'READ'
} as const;
export type UserNotificationDtoStatusEnum = typeof UserNotificationDtoStatusEnum[keyof typeof UserNotificationDtoStatusEnum];

/**
 * @export
 */
export const UserNotificationDtoResultEnum = {
    Success: 'Success',
    Failure: 'Failure'
} as const;
export type UserNotificationDtoResultEnum = typeof UserNotificationDtoResultEnum[keyof typeof UserNotificationDtoResultEnum];